pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-ai-agent"
version = "0.1.0"
description = "Autonomous JARVIS AI Agent - Copilot for coding, research, and automation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "JARVIS AI Team"}]
keywords = ["ai", "agent", "automation", "coding", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
voice = [
    "openai-whisper>=20231117",
    "elevenlabs>=0.2.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "pvporcupine>=2.2.0",
]
ml = [
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[tool.setuptools]
packages = ["jarvis"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
